+++ apps/browser/app/api/auth/[...nextauth]/route.ts
@@ -1,24 +1,5 @@
+import { authOptions } from "@/lib/api/authOptions";
 import NextAuth from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import GithubProvider from "next-auth/providers/github";
-import { SupabaseAdapter } from "@auth/supabase-adapter";
-
-export const authOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-    GithubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID as string,
-      clientSecret: process.env.GITHUB_CLIENT_SECRET as string,
-    }),
-  ],
-  adapter: SupabaseAdapter({
-    url: process.env.SUPABASE_URL as string,
-    secret: process.env.SUPABASE_SERVICE_ROLE_KEY as string
-  })
-};
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
