+++ apps/browser/lib/fonts.ts
@@ -1,0 +1,5 @@
+import { Exo } from "next/font/google";
+
+export const EXO_FONT = Exo({
+  subsets: ["latin"],
+});
